- Formatting - Warning fixes - Rename "test_read_v2_as_v1" to "test_read_v1_as_v2", to match behavior
svn-id: r40726

+++ test/common/bufferedseekablereadstream.h
@@ -22,7 +22,7 @@ class BufferedSeekableReadStreamTestSuite : public CxxTest::TestSuite {
 
 		TS_ASSERT(!ssrs.eos());
 
-		TS_ASSERT_EQUALS(0, ssrs.read(&b, 1));
+		TS_ASSERT_EQUALS((uint)0, ssrs.read(&b, 1));
 		TS_ASSERT(ssrs.eos());
 	}
