+++ test/common/seekablesubreadstream.h
@@ -24,7 +24,7 @@ class SeekableSubReadStreamTestSuite : public CxxTest::TestSuite {
 		}
 
 		TS_ASSERT(!ssrs.eos());
-		TS_ASSERT_EQUALS(0, ssrs.read(&b, 1));
+		TS_ASSERT_EQUALS((uint)0, ssrs.read(&b, 1));
 		TS_ASSERT(ssrs.eos());
 	}
